feat: 차트 UI 개발

See merge request s13-bigdata-dist-sub1/S13P21A301!43

+++ back/src/main/java/com/stock/survive/entity/StockItems.java
@@ -16,7 +16,7 @@ public class StockItems {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "item_no")
-    private Long itemNo;
+    private Integer itemNo;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "item_category_no", nullable = false)
